@@ -1,3 +1,3 @@
 module hello-go
 
-go 1.22.7
+go 1.25.4
