@@ -3,9 +3,8 @@ package main
 import (
 	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
-	"github.com/docker/docker/pkg/tarsum"
+	"github.com/docker/docker/daemon/builder/remotecontext/tarsum"
 )
 
 func main() {
@@ -15,7 +14,7 @@ func main() {
 		os.Exit(1)
 	}
 
-	if _, err = io.Copy(ioutil.Discard, ts); err != nil {
+	if _, err = io.Copy(io.Discard, ts); err != nil {
 		fmt.Println(err)
 		os.Exit(1)
 	}
